% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOItransition.R
\name{AOItransition}
\alias{AOItransition}
\alias{AOItransitions}
\title{calculate transitions for eye-tracking data}
\usage{
AOItransitions(AOI)
}
\arguments{
\item{AOI}{character vector with one entry per fixation, containing AOI's upon 
which was fixated}
}
\description{
function to calculate transitions between AOI's for eye-tracking data
}
\details{
Indicates each time a participant's fixation moves from one AOI to 
  r AOI whereby absence of transition is coded as 0 and transitions are 
  coded as X-Y whereby
  
  X = ID of AOI left
  
  Y = ID of AOI entered
}
\examples{
# The function accepts a single AOI column
  data(SimData )
  
  AOItransitions( SimData$AOI )

}
\author{
Tine van Daal [aut], \email{tine.vandaal@uantwerpen.be}
  San Verhavert [ctb], \email{san.verhavert@uantwerpen.be}
}
