% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileAOI.R
\name{compileAIO}
\alias{compileAIO}
\title{compile AIO columns in one column}
\usage{
compileAIO(data, AOI, labels = NULL)
}
\arguments{
\item{data}{A data frame containing fixation information of an eye tracing 
experiment. Each row indicates a fixation.}

\item{AOI}{A vector containing the name or number of the columns in \code{data} 
containing if the repsective AOI was fixated (1) or not (0).}

\item{labels}{[optional] A vector containing the names of the AOI in the same 
order as the columnnames or -numbers provided to \code{AOI}}
}
\value{
Vector that contains the AIO fixated on for each row in the data frame 
  whereby absence of a fixation on AIO is coded as 0.
}
\description{
Compiles the information on AIO's in separate variables to one variable
}
\details{
This function can be used to convert a wide format eye tracking data 
  frame to a long format eye tracking data frame. It takes a data frame with 
  multiple binary variables that indicate whether an AIO is fixated on (=1) or 
  not (=0) and returns a vector with the AOI's that were fixated on.  
  
  NOTE: if the names of the AOI columns passed to the function are just numbers, 
  make shure to pass these as character for the function to work properly. The 
  function does not check for this case.
}
\author{
Tine van Daal [aut], \email{tine.vandaal@uantwerpen.be}
  San Verhavert [ctb], \email{san.verhavert@uantwerpen.be}
}
