% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{size2deg}
\alias{px2deg}
\alias{deg2size}
\alias{deg2px}
\title{convert between units in eye-tracking}
\usage{
size2deg(x, dist)

px2deg(x, dist, res, screenW)

deg2size(x, dist)

deg2px(x, dist, res, screenW)
}
\arguments{
\item{x}{The value to convert. This can be a single number or a numerical 
vector.}

\item{dist}{The distance to the screen. See Details for the measurement unit.}

\item{res}{A vector indicating the screen resolution in the horizontal 
direction (in pixels).}

\item{screenW}{The width of the screen in the horizontal direction. See 
Details for the measurement unit.}
}
\value{
Returns a value or vector of values indicating the corresponding visual 
  degrees.
}
\description{
The functions convert  between visual degrees and centimeters, inches or pixels.
}
\details{
When converting from size to degrees (\code{size2deg}), the measurement 
  unit of the distance to the screen (\code{dist}) should be consistent with 
  the measurement unit of the size you are converting rf--from (\code{x}. 
  Namely if the size is in centimeters, then the distance should be to, and if 
  the size is in inches, then the distance should be in inches.

When converting from pixels to degrees (\code{px2deg}), the measurement 
  unit of the distance to the screen (\code{dist}) should be consistent with 
  the measurement unit of the screen width (\code{screen width}). Namely if 
  the size is in centimeters, then the screen width should be to, and if the 
  size is in inches, then the screen width should be in inches.
}
\section{Functions}{
\itemize{
\item \code{px2deg}: 

\item \code{deg2size}: 

\item \code{deg2px}: 
}}

