% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDur.R
\name{fixDur}
\alias{fixDur}
\title{Fixation Duration}
\usage{
fixDur(data, fixTime, passes)
}
\arguments{
\item{data}{A data frame containing fixation information of an eye tracing 
experiment and the coded passes. Each row indicates a fixation.}

\item{fixTime}{The name or number of the column containing the time per fixation.}

\item{passes}{The name or number of the column containing the coded passes.}
}
\value{
A vector containing the agregated fixation duration for the passes or 
  AOI's, dependent upon what is provided to \code{passes}. The result will be 
  in the same unit as the duration input.
}
\description{
Calculates the fixation durations for the first and second passes
}
\details{
This function is a wrapper for  \code{\link[stats]{aggregate}} 
  
  The function will only return the fixation duration of the values in the 
  \code{passes} column. The passes column is the vector returned by the 
  \code{link{codePasses}} function. It is also possible to provide the column 
  name of the column containing the AOI's if you require the fixation durations 
  for the AOI's only.
  The column of which the name or number is passed to \code{passes},will be 
  converted to a factor if it is not yet the case.
}
\seealso{
\code{\link[stats]{aggregate}}, \code{\link[base]{by}}, 
  \code{\link[base]{tapply}}
}
